Update gpt, gpt2, opt
